requirements.txt
streamlit
streamlit-lottie
pillow
rembg
